Testing GHCJS web target
