+++ package.yaml
@@ -25,7 +25,7 @@ dependencies:
 - base >= 4.7 && < 5
 - BNFC >= 2.8.1 && <= 2.8.2
 - array >= 0.5.3.0
-- containers >= 0.6.0.1
+- containers >= 0.5.7.1
 - mtl >= 2.2.0.1
 - optparse-applicative >= 0.14.0.0
 - raw-strings-qq >= 1.1
