+++ web/helper.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+/root/IdeaProjects/ocaml.hs/web/.stack-work/dist/x86_64-linux/Cabal-1.24.0.0_ghcjs/build/reflex-starter/reflex-starter.jsexe/
